requirements.txt
#-i https://pypi.tuna.tsinghua.edu.cn/simple
torch==2.3.1
torchaudio==2.3.1
torchelastic==0.2.2
torchvision==0.18.1
nibabel
numpy
scipy
antspyx
